[MNOE-783] Fix registration route not being loaded
Fix an issue when the registration is disabled in the settings.yml and
then enabled via the MnoHub settings.
Since the routes are being build based on the User#devise_modules which
is built when loading the class, the registration route is never added.

The Devise::Models::Registerable is only adding a class method to the
user model (.new_with_session), so it's ok to always add it.